refactor: edit description and file structure
