+++ multisig/simulations/sepolia/index.ts
@@ -10,3 +10,4 @@ import "./vip-009";
 // import './vip-010' // passes when run separately
 import "./vip-013";
 import "./vip-016";
+import "./vip-019";
